Do not include paragraph with modal opener for those with a custom cookie policy
